+++ lib/CMakeLists.txt
@@ -12,7 +12,6 @@ add_subdirectory(Markup)
 add_subdirectory(Option)
 add_subdirectory(Parse)
 add_subdirectory(PrintAsObjC)
-add_subdirectory(Reflection)
 add_subdirectory(Sema)
 add_subdirectory(Serialization)
 if(CMAKE_SYSTEM_NAME STREQUAL "Darwin")
