--- lib/Reflection/CMakeLists.txt
@@ -1,3 +1,0 @@
-add_swift_library(swiftReflection
-  TypeRef.cpp
-  LINK_LIBRARIES swiftBasic)
