+++ stdlib/public/Reflection/CMakeLists.txt
@@ -1,0 +1,5 @@
+add_swift_library(swiftReflection IS_STDLIB
+  Demangle.cpp
+  Remangle.cpp
+  TypeRef.cpp
+  INSTALL_IN_COMPONENT stdlib)
