+++ stdlib/public/Reflection/Demangle.cpp
@@ -1,0 +1,3 @@
+#include "../../../lib/Basic/Demangle.cpp"
+#include "../../../lib/Basic/Punycode.cpp"
+#include "../../../lib/Basic/PunycodeUTF8.cpp"
