+++ test/lit.cfg
@@ -272,6 +272,7 @@ config.sil_extract = inferSwiftBinary('sil-extract')
 config.lldb_moduleimport_test = inferSwiftBinary('lldb-moduleimport-test')
 config.swift_ide_test = inferSwiftBinary('swift-ide-test')
 config.swift_reflection_test = inferSwiftBinary('swift-reflection-test')
+config.swift_reflection_dump = inferSwiftBinary('swift-reflection-dump')
 config.clang = inferSwiftBinary('clang')
 config.llvm_link = inferSwiftBinary('llvm-link')
 config.swift_llvm_opt = inferSwiftBinary('swift-llvm-opt')
@@ -840,7 +841,8 @@ config.substitutions.append(
      (subst_target_swift_ide_test_mock_sdk,
       subst_target_swift_ide_test_mock_sdk_after)))
 config.substitutions.append(('%target-swift-ide-test', config.target_swift_ide_test))
-config.substitutions.append(('%target-swift-reflection-test', '{} {} {}'.format(config.swift_reflection_test, '-arch', run_cpu)))
+config.substitutions.append(('%target-swift-reflection-test', '{test_runner}{variant_suffix} {arch}'.format(test_runner=config.swift_reflection_test, variant_suffix=config.variant_suffix, arch=run_cpu)))
+config.substitutions.append(('%target-swift-reflection-dump', '{} {} {}'.format(config.swift_reflection_dump, '-arch', run_cpu)))
 config.substitutions.append(('%target-swiftc_driver', config.target_swiftc_driver))
 
 if hasattr(config, 'target_swift_autolink_extract'):
