+++ test/lit.site.cfg.in
@@ -15,6 +15,7 @@ config.target_triple = "@TARGET_TRIPLE@"
 config.targets_to_build = "@TARGETS_TO_BUILD@"
 config.variant_triple = "@VARIANT_TRIPLE@"
 config.variant_sdk = "@VARIANT_SDK@"
+config.variant_suffix = "@VARIANT_SUFFIX@"
 config.swiftlib_dir = "@LIT_SWIFTLIB_DIR@"
 config.darwin_xcrun_toolchain = "@SWIFT_DARWIN_XCRUN_TOOLCHAIN@"
