+++ tools/CMakeLists.txt
@@ -5,7 +5,6 @@ add_subdirectory(swift-demangle)
 add_subdirectory(lldb-moduleimport-test)
 add_subdirectory(sil-extract)
 add_subdirectory(swift-llvm-opt)
-add_subdirectory(swift-reflection-test)
 
 if(SWIFT_BUILD_SOURCEKIT)
   add_subdirectory(SourceKit)
